pyproject.toml
[tool.poetry]
name = "sasrec"
version = "0.1.0"
description = "Self-Attentive Sequential Recommendation model(SASRec)"
authors = [
    "Venkata Harshit Koneru <koneru.v@zdf.de>",
    "Xenija Neufeld <neufeld.x@zdf.de>",
    "Christian Malan <malan.c@zdf.de>",
    "Kuhn, Valentin <kuhn.v@zdf.de>",
    "Netzer, Florian <netzer.f@zdf.de>",
    "Meyer, James <meyer.jam@zdf.de>",
]
maintainers = ["Venkata Harshit Koneru <koneru.v@zdf.de>"]
readme = "README.md"
packages = [{ include = "sasrec" }]

[tool.poetry.dependencies]
python = "~3.8.1"
# install pip, setuptools, wheel before installing scikit-surprise, lightfm, recommenders to fix build errors in the latter
# pip = "^23.3.2"
# setuptools = "^69.0.3"
# wheel = "^0.42.0"
pa-base = { version = "~3.0.0"}
mlflow = "^2.8.0"
tensorflow-io-gcs-filesystem = "0.27.0"
scikit-surprise = "1.1.3"
lightfm = "1.17"
# recommenders only supports python = ">=3.6,<3.10"
recommenders = { version = "1.1.1", extras = ["gpu"] }
# install recommenders from git source which is ~1.5 years more up to date than latest version 1.1.1
# recommenders = { git = "https://github.com/recommenders-team/recommenders.git", rev = "0d9d7c7", extras = [
#     "gpu",
# ] }
Flask="~2.0.1"
protobuf="~3.19.6"
Werkzeug="~2.2.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pre-commit = "^3.3.3"
pytest = "^7.4.0"
python-dotenv = "^1.0.0"

[build-system]
requires = [
    "poetry-core>=1.0.0",
    # add all the below packages (pip, setuptools, wheel, importlib-metadata, urllib3, idna, certifi, packaging) to fix build errors in scikit-surprise, lightfm, recommenders
    # "pip>=23.3.2",
    # "setuptools>=69.0.3",
    # "wheel>=0.42.0",
    # "importlib-metadata",
    # "urllib3",
    # "idna",
    # "certifi",
    # "packaging",
]
build-backend = "poetry.core.masonry.api"
